% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{chf_NTS}
\alias{chf_NTS}
\title{chf_NTS}
\usage{
chf_NTS(u, param)
}
\arguments{
\item{u}{An array of u}

\item{ntsparam}{A vector of the NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
For NTS process case it is a vector of parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu, t)}.}
}
\value{
Characteristic function of the NTS distribution
}
\description{
\code{chf_NTS} calculates Ch.F of the NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
If a time parameter value is given, it calculates Ch.F of the NTS profess
\eqn{\phi(u)=E[\exp(i u (X(t+s)-X(s)) )]=\exp(t \log(E[\exp(i u X(1))]))}, where X is the NTS process generated
by the NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
}
\examples{
library("temStaR")
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
u <- seq(from = -2*pi, to = 2*pi, length.out = 101)
phi <- chf_NTS(u, ntsparam)


#Annual based parameters
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
#scaling annual parameters to one day
dt <- 1/250 #one day
ntsparam <- c(alpha, theta, beta, gamma, mu, dt)
u <- seq(from = -2*pi, to = 2*pi, length.out = 101)
phi <- chf_NTS(u, ntsparam)

}
