% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{chf_stdNTS}
\alias{chf_stdNTS}
\title{chf_stdNTS}
\usage{
chf_stdNTS(u, param)
}
\arguments{
\item{u}{An array of u}

\item{ntsparam}{A vector of the standard NTS parameters \eqn{(\alpha, \theta, \beta)}.
For the standard NTS process case it is a vector of parameters \eqn{(\alpha, \theta, \beta, t)}.}
}
\value{
Characteristic function of the standatd NTS distribution
}
\description{
\code{chf_stdNTS} calculates Ch.F of the standard NTS distribution with parameters \eqn{(\alpha, \theta, \beta)}.
If a time parameter value is given, it calculates Ch.F of the standard NTS profess
\eqn{\phi(u)=E[\exp(i u (X(t+s)-X(s))]=\exp(t \log(E[\exp(i u X(1))]))}, where X is the standard NTS process generated
by the standard NTS distribution with parameters \eqn{(\alpha, \theta, \beta)}.
}
\examples{
library("temStaR")
alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
u <- seq(from = -2*pi, to = 2*pi, length.out = 101)
phi <- chf_stdNTS(u, ntsparam)


#Annual based parameters
alpha <- 1.2
theta <- 1
beta <- -0.2
#scaling annual parameters to one day
dt <- 1/250 #one day
ntsparam <- c(alpha, theta, beta, gamma, mu, dt)
u <- seq(from = -2*pi, to = 2*pi, length.out = 101)
phi <- chf_stdNTS(u, ntsparam)

}
