% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{cvarmarginalmnts}
\alias{cvarmarginalmnts}
\title{cvarmarginalmnts}
\usage{
cvarmarginalmnts(eta, n, st)
}
\arguments{
\item{eta}{the significant level for CVaR. Real value between 0 and 1.}

\item{n}{the \eqn{n}-th element to be calculated.}

\item{st}{Structure of parameters for the n-dimensional NTS distribution.}
}
\description{
\code{cvarmarginalmnts} calculates
the CVaR of the \eqn{n}-th element
of the multivariate NTS distributed random variable.
}
