% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{dcopulaStdNTS}
\alias{dcopulaStdNTS}
\title{dcopulaStdNTS}
\usage{
dcopulaStdNTS(u, st, subTS = NULL)
}
\description{
\code{dcopulaStdNTS} calculates
density of the stdNTS copula.
}
\references{
Y. S. Kim, D. Volkmann (2013), Normal Tempered Stable Copula, Applied Mathematics Letters,  26(7), 676-680
\url{https://www.sciencedirect.com/science/article/pii/S0893965913000384}
}
