% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{dCVaR_numint}
\alias{dCVaR_numint}
\title{dCVaR_numint}
\usage{
dCVaR_numint(eta, alpha, theta, beta, N = 200, rho = 0.1)
}
\description{
The first derivative of CVaR for the beta parameter of the stdNTS.
Developer's version.
}
