% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{dCVaRstdNTS_numint}
\alias{dCVaRstdNTS_numint}
\title{dCVaRstdNTS_numint}
\usage{
dCVaRstdNTS_numint(
  eta,
  alpha,
  theta,
  beta,
  cv = NULL,
  v = NULL,
  N = 20,
  rho = 1e-04
)
}
\description{
Calculate the marginal contribution to CVaR for the multivariate stdNTS Model.
Developer's version.
}
