% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{dinvCdf_stdNTS}
\alias{dinvCdf_stdNTS}
\title{dinvCdf_stdNTS}
\usage{
dinvCdf_stdNTS(eta, alpha, theta, beta)
}
\description{
The first derivative of inverse CDF for the beta parameter of the stdNTS.
Developer's version.
}
