% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{dinvCdf_stdNTS_int}
\alias{dinvCdf_stdNTS_int}
\title{dinvCdf_stdNTS_int}
\usage{
dinvCdf_stdNTS_int(eta, x = NULL, alpha, theta, beta)
}
\description{
The first derivative of inverse CDF for the beta parameter of the stdNTS.
Developer's version.
}
