% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{dmarginalmnts}
\alias{dmarginalmnts}
\title{dmarginalmnts}
\usage{
dmarginalmnts(x, n, st)
}
\arguments{
\item{x}{the \eqn{x} such that \eqn{f(x) = \frac{d}{dx}P(X_n<x)}}

\item{n}{the \eqn{n}-th element to be calculated.}

\item{st}{Structure of parameters for the n-dimensional NTS distribution.}
}
\description{
\code{dmarginalmnts} calculates
the marginal density of the \eqn{n}-th element
of the multivariate NTS distributed random variable.
}
