% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{dmnts}
\alias{dmnts}
\title{dmnts}
\usage{
dmnts(x, st, subTS = NULL)
}
\arguments{
\item{x}{array of the \eqn{(x_1, \cdots, x_n)}}

\item{st}{Structure of parameters for the n-dimensional NTS distribution.

\code{st$ndim} : dimension

\code{st$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{st$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{st$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{st$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{st$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{st$Rho} : \eqn{\rho} matrix of the std NTS distribution (X).}

\item{numofsample}{number of samples.}
}
\value{
Simulated NTS random vectors
}
\description{
\code{dmnts} calculates the density of the multivariate NTS distribution:
\eqn{f(x_1, \cdots, x_n)=\frac{d^n}{dx_1\cdots dx_n}P(x_n<R_1, \cdots, x_n<R_n)}.
The multivariate NTS random vector \eqn{R = (R_1, \cdots, R_n)} is defined

\eqn{R = \mu + diag(\sigma) X},

where

\eqn{X} follows \eqn{stdNTS_n(\alpha, \theta, \beta, \Sigma)}
}
\examples{
library("temStaR")
library(mvtnorm)
strPMNTS <- list(ndim = 2,
              mu = c( 0.5, -1.5 ),
              sigma = c( 2, 3 ),
              alpha = 0.1,
              theta = 3,
              beta =  c( 0.1, -0.3 ),
              Rho = matrix( data = c(1.0, 0.75, 0.75, 1.0),
                            nrow = 2, ncol = 2)
)
dmnts(c(0.6, -1.0), st = strPMNTS)


strPMNTS <- list(ndim = 2,
                 mu = c( 0, 0, 0 ),
                 sigma = c( 1, 1, 1 ),
                 alpha = 0.1,
                 theta = 3,
                 beta =  c( 0.1, -0.3, 0 ),
                 Rho = matrix(
                     data = c(1.0, 0.75, 0.1, 0.75, 1.0, 0.2, 0.1, 0.2, 1.0),
                     nrow = 3, ncol = 3)
)
pmnts(c(0,0,0), st = strPMNTS)
dmnts(c(0,0,0), st = strPMNTS)

}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
