% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{fitmnts_par}
\alias{fitmnts_par}
\title{fitmnts}
\usage{
fitmnts_par(
  returndata,
  n,
  alphaNtheta = NULL,
  stdflag = FALSE,
  parallelSocketCluster = NULL
)
}
\description{
\code{fitmnts} fit parameters
of the n-dimensional NTS distribution. A parallel version of fitmnts()
}
\examples{
library(functional)
library(nloptr)
library(pracma)
library(spatstat)
library(Matrix)
library(foreach)
library(doParallel)
library(quantmod)

getSymbols("^GSPC", src="yahoo", from = "2016-1-1", to = "2020-08-31")
pr1 <- as.numeric(GSPC$GSPC.Adjusted)
getSymbols("^DJI", src="yahoo", from = "2016-1-1", to = "2020-08-31")
pr2 <- as.numeric(DJI$DJI.Adjusted)

returndata <- matrix(data = c(diff(log(pr1)),diff(log(pr2))),
                     ncol = 2, nrow = (length(pr1)-1))

numofcluster  <- detectCores()
cl <- makePSOCKcluster(numofcluster)
registerDoParallel(cl)
res <- fitmnts_par( returndata = returndata, n=2, parallelSocketCluster = cl )
stopCluster(cl)

}
