% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{fitnts}
\alias{fitnts}
\title{fitnts}
\usage{
\code{fitnts(rawdat)}
\code{fitnts(rawdat), ksdensityflag = 1}
\code{fitnts(rawdat, initialparam = c(alpha, theta, beta, gamma, mu))}
\code{fitnts(rawdat, initialparam = c(alpha, theta, beta, gamma, mu)), ksdensityflag = 1}
\code{fitnts(rawdat, initialparam = c(alpha, theta, beta, gamma, mu)), maxeval = 100, ksdensityflag = 1}
}
\arguments{
\item{rawdat}{Raw data to fit the parameters.}

\item{initialparam}{A vector of initial NTS parameters.
This function uses the \code{nloptr} package.
If it has a good initial parameter then estimation performs better.
If users do not know a good initial parameters, then just set it as \code{initialparam=NaN}, that is default.
The function \code{cffitnts()} may be helpful to find the initial parameters.}

\item{maxeval}{Maximum evaluation number for \code{nloptr}. The iteration stops on this many function evaluations.}

\item{ksdensityflag}{This function fit the parameters using the curvefit method between the empirical cdf and the NTS cdf.
If \code{ksdensityflag = 1} (default), then the empirical cdf is calculated by the kernel density estimation.
If \code{ksdensityflag = 0}, then the empirical cdf is calculated by the empirical cdf.}
}
\value{
Estimated parameters
}
\description{
\code{fitnts} fit parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)} of the NTS distribution.
This function using the curvefit method between the empirical cdf and the NTS cdf.
}
\examples{
library(evmix)
library("temStaR")
library("quantmod")
getSymbols("^GSPC", src="yahoo", from = "2013-1-1", to = "2023-12-31")
pr <- as.numeric(GSPC$GSPC.Adjusted)
ret <- diff(log(pr))
ntsparam <-  fitnts(ret)

Femp = ecdf(ret)
x = seq(from=min(ret), to = max(ret), length.out = 100)
cemp = Femp(x)
ncdf = pnts(x, c(ntsparam))
plot(x,ncdf,type = 'l', col = "red")
points(x,cemp, type = 'l', col = "blue")
a = density(ret)
p = dnts(x,ntsparam)
plot(x,p,type = 'l', col = "red")
lines(a,type = 'l', col = "blue")

}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
