% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{fitstdntsFixAlphaThata}
\alias{fitstdntsFixAlphaThata}
\title{fitstdntsFixAlphaThata}
\usage{
fitstdntsFixAlphaThata(
  rawdat,
  alpha,
  theta,
  initialparam = NaN,
  maxeval = 100,
  ksdensityflag = 1
)
}
\description{
Fit beta of stdNTS distribution with fixed alpha and theta.
}
