% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{gensamplepathnts}
\alias{gensamplepathnts}
\title{gensamplepathnts}
\usage{
gensamplepathnts(npath, ntimestep, ntsparam, dt)
}
\arguments{
\item{npath}{Number of sample paths}

\item{ntimestep}{number of time step}

\item{ntsparam}{A vector of the NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
A vector of the standard NTS parameters \eqn{(\alpha, \theta, \beta)}.}

\item{dt}{the time length of one time step by the year fraction. "dt=1" means 1-year.}
}
\value{
Structure of the sample path.
Matrix of sample path. Column index is time.
}
\description{
\code{gensamplepathnts} generate sample paths
of the NTS process with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
If only three parameters are given, it generate sample paths
of the standard NTS process with parameters \eqn{(\alpha, \theta, \beta)}.
}
\examples{
library("temStaR")
#standard NTS process sample path
alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
npath <- 5
ntimestep <- 250
dt <- 1/250
simulation <- gensamplepathnts(npath, ntimestep, ntsparam, dt)
matplot(colnames(simulation), t(simulation), type = 'l')

#NTS process sample path
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
npath <- 5
ntimestep <- 250
dt <- 1/250
simulation <- gensamplepathnts(npath, ntimestep, ntsparam, dt)
matplot(colnames(simulation), t(simulation), type = 'l')

}
