% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{getPortNTSParam}
\alias{getPortNTSParam}
\title{getPortNTSParam}
\usage{
\code{res <- setPortfolioParam(strPMNTS,w)}
\code{res <- setPortfolioParam(strPMNTS,w, FALSE)}
}
\arguments{
\item{strPMNTS}{Structure of parameters for the n-dimensional NTS distribution.

\code{strPMNTS$ndim} : dimension

\code{strPMNTS$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{strPMNTS$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{strPMNTS$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{strPMNTS$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{strPMNTS$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{res$Rho} : \eqn{\rho} matrix (Correlation) of the std NTS distribution (X).

\code{res$Sigma} : Covariance \eqn{\Sigma} matrix of return data \eqn{r}.}

\item{w}{Capital allocation weight vector.}

\item{stdform}{If \code{stdform} is \code{FALSE}, then the return parameter has the following representation

\eqn{R_p = <w, r> = \mu + diag(\sigma) X},

where

\eqn{X} follows \eqn{stdNTS_1(\alpha, \theta, \beta, 1)}.

If \code{stdform} is \code{TRUE}, then the return parameter has the following representation

\eqn{R_p = <w, r>} follows \eqn{NTS_1(\alpha, \theta, \beta, \gamma, \mu, 1)}}
}
\value{
The weighted sum follows 1-dimensional NTS.

\eqn{R_p = <w, r> = \mu + diag(\sigma) X},

where

\eqn{X} follows \eqn{stdNTS_1(\alpha, \theta, \beta, 1)}.

Hence we obtain

\code{res$mu} : \eqn{\mu} mean of \eqn{R_p}.

\code{res$sigma} : \eqn{\sigma} standard deviation of \eqn{R_p}.

\code{res$alpha} : \eqn{\alpha} of \eqn{X}.

\code{res$theta} : \eqn{\theta} of \eqn{X}.

\code{res$beta} : \eqn{\beta} of \eqn{X}.
}
\description{
Portfolio return with capital allocation weight is \eqn{R_p=<w,r>},
which is a weighted sum of of elements in the N-dimensional NTS random vector.
\eqn{R_p} becomes an 1-dimensional NTS random variable.
\code{getPortNTSParam} find the parameters of \eqn{R_p}.
}
\examples{
library("temStaR")
strPMNTS <- list(ndim = 2,
                 mu = c( 9.876552e-05, 4.747343e-04 ),
                 sigma = c( 0.01620588, 0.02309643 ),
                 alpha = 0.1888129 ,
                 theta = 0.523042,
                 beta =  c( -0.04632938,  0.04063555 ),
                 Rho = matrix( data = c(1.0, 0.469883,
                                       0.469883, 1.0),
                               nrow = 2, ncol = 2)
                 CovMtx = matrix( data = c(0.0002626304, 0.0001740779,
                                         0.0001740779, 0.0005334452),
                                 nrow = 2, ncol = 2)
                 )
w <- c(0.3, 0.7)
res <- getPortNTSParam(strPMNTS,w)

}
\references{
Proposition 2.1 of
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
