% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{importantSamplining}
\alias{importantSamplining}
\title{importantSamplining}
\usage{
importantSamplining(alpha, theta)
}
\description{
\code{importantSamplining} do the important sampling for the TS Subordinator.
}
