% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{mctCVaRnts}
\alias{mctCVaRnts}
\title{mctCVaRnts}
\usage{
mctCVaRnts(
  eta,
  n,
  w,
  covMtx,
  alpha,
  theta,
  betaArray,
  muArray,
  CVaR = NULL,
  dCVaR = NULL
)
}
\description{
Calculate the marginal contribution to CVaR for the multivariate NTS market model.
Developer's version.
}
