% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{mctStdDev}
\alias{mctStdDev}
\title{mctStdDev}
\usage{
mctStdDev(n, w, covMtx)
}
\arguments{
\item{n}{The targer stock to calculate the mctStdDev}

\item{w}{The capital allocation rate vector for the current portfolio}

\item{CovMtx}{Covariance matrix of return data.}
}
\description{
Morginal contribution to Risk for Standard Deviation.
}
