% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{mctVaRmnts}
\alias{mctVaRmnts}
\title{mctVaRmnts}
\usage{
\code{mctVaRmnts(eta, n, w, st)}
}
\arguments{
\item{eta}{Significant level of CVaR.}

\item{n}{The targer stock to calculate the mctCVaR}

\item{w}{The capital allocation rate vector for the current portfolio}

\item{st}{Structure of parameters for the N-dimensional NTS distribution.

\code{st$ndim} : Dimension of the model. Here \code{st$ndim=N}.

\code{st$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{st$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{st$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{st$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{st$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{st$Rho} : \eqn{\rho} matrix of the std NTS distribution (X),
which is correlation matrix of epsilon.

\code{st$CovMtx} : Covariance matrix of return data \eqn{r}.}
}
\description{
Calculate the marginal contribution to VaR for the multivariate NTS market model:
the random vector \eqn{r} is

\eqn{r = \mu + diag(\sigma) X}

where

\eqn{X} follows \eqn{stdNTS_N(\alpha, \theta, \beta, \Sigma)}
}
\examples{
library(functional)
library(nloptr)
library(pracma)
library(spatstat)
library(Matrix)
library(quantmod)
library(mvtnorm)
library("temStaR")

#Fix alpha and theta.
#Estimate alpha dna theta from DJIA and use those parameter for IBM, INTL parameter fit.
getSymbols("^DJI", src="yahoo", from = "2020-8-25", to = "2020-08-31")
prDJ <- as.numeric(DJI$DJI.Adjusted)
ret <- diff(log(prDJ))
ntsparam <-  fitnts(ret)
getSymbols("IBM", src="yahoo", from = "2016-1-1", to = "2020-08-31")
pr1 <- as.numeric(IBM$IBM.Adjusted)
getSymbols("INTL", src="yahoo", from = "2016-1-1", to = "2020-08-31")
pr2 <- as.numeric(INTL$INTL.Adjusted)

returndata <- matrix(data = c(diff(log(pr1)),diff(log(pr2))),
                     ncol = 2, nrow = (length(pr1)-1))
st <- fitmnts( returndata = returndata,
                n = 2,
                alphaNtheta = c(ntsparam["alpha"], ntsparam["theta"])  )
w <- c(0.3, 0.7)
eta <- 0.01

mctVaRmnts(eta, 1, w, st) #MCT-VaR for IBM
mctVaRmnts(eta, 2, w, st) #MCT-VaR for INTL

mctCVaRmnts(eta, 1, w, st) #MCT-CVaR for IBM
mctCVaRmnts(eta, 2, w, st) #MCT-CVaR for INTL

}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
