% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{mctVaRnts}
\alias{mctVaRnts}
\title{mctVaRnts}
\usage{
mctVaRnts(
  eta,
  n,
  w,
  covMtx,
  alpha,
  theta,
  betaArray,
  muArray,
  icdf = NULL,
  dicdf = NULL
)
}
\description{
Calculate the marginal contribution to VaR for the multivariate NTS market model.
Developer's version.
}
