% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{moments_NTS}
\alias{moments_NTS}
\title{moments_NTS}
\usage{
moments_NTS(param)
}
\arguments{
\item{param}{A vector of the NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.}
}
\value{
First 4 moments (Mean, Variance, Skewness, Excess Kurtosis) of NTS distribution.
The mean is always the same as the parameter \eqn{\mu}.
}
\description{
\code{moments_NTS} calculates mean, variance, skewness, and excess kurtosis
of the NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
}
\examples{
library("temStaR")
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
moments_NTS(param = ntsparam)
}
\references{
Kim, Y.S, K-H Roh, R. Douady (2020) Tempered Stable Processes with Time Varying Exponential Tails
\url{https://arxiv.org/pdf/2006.07669.pdf}
}
