% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{moments_stdNTS}
\alias{moments_stdNTS}
\title{moments_stdNTS}
\usage{
moments_stdNTS(param)
}
\arguments{
\item{param}{A vector of the standard NTS parameters \eqn{(\alpha, \theta, \beta)}.}
}
\value{
First 4 moments (Mean, Variance, Skewness, Excess Kurtosis) of NTS distribution.
Of course, the mean and variance are always 0 and 1, respectively.
}
\description{
\code{moments_stdNTS} calculates mean, variance, skewness, and excess kurtosis
of the standard NTS distribution with parameters \eqn{(\alpha, \theta, \beta)}.
}
\examples{
library("temStaR")
alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
moments_stdNTS(param = ntsparam)
}
\references{
Kim, Y.S, K-H Roh, R. Douady (2020) Tempered Stable Processes with Time Varying Exponential Tails
\url{https://arxiv.org/pdf/2006.07669.pdf}
}
