% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{pmarginalmnts}
\alias{pmarginalmnts}
\title{pmarginalmnts}
\usage{
pmarginalmnts(x, n, st)
}
\arguments{
\item{x}{the \eqn{x} such that \eqn{F(x) = P(X_n<x)}}

\item{n}{the \eqn{n}-th element to be calculated.}

\item{st}{Structure of parameters for the n-dimensional NTS distribution.}
}
\description{
\code{pmarginalmnts} calculates
the marginal cdf of the \eqn{n}-th element
of the multivariate NTS distributed random variable.
}
