% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaRMultiNTS.R
\name{portfolioVaRETmnts}
\alias{portfolioVaRETmnts}
\title{portfolioVaRETmnts}
\usage{
portfolioVaRETmnts(strPMNTS, w, eta)
}
\arguments{
\item{strPMNTS}{Structure of parameters for the n-dimensional NTS distribution.

\code{strPMNTS$ndim} : dimension

\code{strPMNTS$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{strPMNTS$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{strPMNTS$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{strPMNTS$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{strPMNTS$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{res$Rho} : \eqn{\rho} matrix (Correlation) of the std NTS distribution (X).

\code{res$Sigma} : Covariance \eqn{\Sigma} matrix of return data \eqn{r}.}

\item{w}{Capital allocation weight vector.}

\item{eta}{significanlt level}
}
\value{
portfolio value at Return on the NTS market model
}
\description{
Calculate portfolio value at return on the NTS market model
}
