% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{qnts}
\alias{qnts}
\title{qnts}
\usage{
qnts(u, ntsparam)
}
\arguments{
\item{u}{vector of probabilities.}

\item{ntsparam}{A vector of the NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
For the NTS process case it is a vector of parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu, t)}.
A vector of standard NTS parameters \eqn{(\alpha, \theta, \beta)}.}
}
\value{
The quantile function of the NTS distribution
}
\description{
\code{qnts} calculates quantile of the NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
If only three parameters are given, it calculates quantile of the standard NTS distribution with parameter \eqn{(\alpha, \theta, \beta)}
If a time parameter value is given, it calculates quantile of NTS profess. That is it finds x such that
\eqn{u = P((X(t+s)-X(s))<x)}, where X is the NTS process generated
by the NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
}
\examples{
library("temStaR")

alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
u <- c(0.01,0.05,0.25,0.5, 0.75, 0.95, 0.99)
q <- qnts(u, ntsparam)

alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
u <- c(0.01,0.05,0.25,0.5, 0.75, 0.95, 0.99)
q <- qnts(u, ntsparam)


#Annual based parameters
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
#scaling annual parameters to one day
dt <- 1/250 #one day
ntsparam <- c(alpha, theta, beta, gamma, mu, dt)
u <- c(0.01,0.05,0.25,0.5, 0.75, 0.95, 0.99)
q <- qnts(u, ntsparam)
}
