% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{rmnts}
\alias{rmnts}
\title{rmnts}
\usage{
rmnts(strMnts, numofsample, u = NULL)
}
\arguments{
\item{numofsample}{number of samples.}

\item{strPMNTS}{Structure of parameters for the n-dimensional NTS distribution.

\code{strPMNTS$ndim} : dimension

\code{strPMNTS$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{strPMNTS$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{strPMNTS$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{strPMNTS$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{strPMNTS$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{strPMNTS$Rho} : \eqn{\rho} matrix of the std NTS distribution (X).}
}
\value{
Simulated NTS random vectors
}
\description{
\code{rmnts} generates random vector following the n dimensional NTS distribution using Cholesky decomposition.

\eqn{r = \mu + diag(\sigma) X},

where

\eqn{X} follows \eqn{stdNTS_n(\alpha, \theta, \beta, \Sigma)}
}
\examples{
strPMNTS <- list(ndim = 2,
                 mu = c( 0.00011, 0.00048 ),
                 sigma = c( 0.0162, 0.0231 ),
                 alpha = 1.23,
                 theta = 3.607,
                 beta =  c( -0.1209,  0.0905 ),
                 Rho = matrix( data = c(1.0, 0.55, 0.55, 1.0), nrow = 2, ncol = 2)
)
gensim <- rmnts( strPMNTS, 100 )
plot(gensim)

}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
