% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{rnts}
\alias{rnts}
\title{rnts}
\usage{
rnts(n, ntsparam)
}
\arguments{
\item{n}{number of random numbers to be generated.}

\item{ntsparam}{A vector of NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
For NTS process case it is a vector of parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu, t)}.
A vector of standard NTS parameters \eqn{(\alpha, \theta, \beta)}.}
}
\value{
NTS random numbers
}
\description{
\code{rnts} generates random numbers following NTS distribution with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
If only three parameters are given, it generates random numbers of standard NTS distribution with parameter \eqn{(\alpha, \theta, \beta)}
If a time parameter value is given, it generates random numbers of increments of NTS profess for time interval t.
}
\examples{
library("temStaR")

alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
r <- rnts(100, ntsparam) #generate 100 NTS random numbers
plot(r)

alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
r <- rnts(100, ntsparam) #generate 100 NTS random numbers
plot(r)


#Annual based parameters
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
#scaling annual parameters to one day
dt <- 1/250 #one day
ntsparam <- c(alpha, theta, beta, gamma, mu, dt)
r <- rnts(100, ntsparam) #generate 100 NTS random numbers
plot(r)
}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
