% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{setPortfolioParam}
\alias{setPortfolioParam}
\title{setPortfolioParam}
\usage{
\code{res <- setPortfolioParam(strPMNTS,w)}
}
\arguments{
\item{strPMNTS}{Structure of parameters for the n-dimensional NTS distribution.

\code{strPMNTS$ndim} : dimension

\code{strPMNTS$mu} : \eqn{\mu} mean vector (column vector) of the input data.

\code{strPMNTS$sigma} : \eqn{\sigma} standard deviation vector (column vector) of the input data.

\code{strPMNTS$alpha} : \eqn{\alpha} of the std NTS distribution (X).

\code{strPMNTS$theta} : \eqn{\theta} of the std NTS distribution (X).

\code{strPMNTS$beta} : \eqn{\beta} vector (column vector) of the std NTS distribution (X).

\code{strPMNTS$Rho} : \eqn{\Sigma} matrix of the std NTS distribution (X).}

\item{w}{Capital allocation weight vector.}
}
\value{
The weighted sum follows 1-dimensional NTS.

\eqn{R_p = <w, r> = \mu + diag(\sigma) X},

where

\eqn{X} follows \eqn{stdNTS_1(\alpha, \theta, \beta, 1)}.

Hence we obtain

\code{res$mu} : \eqn{\mu} mean of \eqn{R_p}.

\code{res$sigma} : \eqn{\sigma} standard deviation of \eqn{R_p}.

\code{res$alpha} : \eqn{\alpha} of \eqn{X}.

\code{res$theta} : \eqn{\theta} of \eqn{X}.

\code{res$beta} : \eqn{\beta} \eqn{X}.
}
\description{
Portfolio return with capital allocation weight is \eqn{R_p=<w,r>},
which is a weighted sum of of elements in the N-dimensional NTS random vector.
\eqn{R_p} becomes an 1-dimensional NTS random variable.
\code{setPortfolioParam} find the parameters of \eqn{R_p}.
}
\examples{
strPMNTS <- list(ndim = 2,
                 mu = c( 0.00011, 0.00048 ),
                 sigma = c( 0.0162, 0.0231 ),
                 alpha = 1.23,
                 theta = 3.607,
                 beta =  c( -0.1209,  0.0905 ),
                 Rho = matrix( data = c(1.0, 0.55, 0.55, 1.0), nrow = 2, ncol = 2)
)
w <- c(0.3, 0.7)
res <- setPortfolioParam(strPMNTS,w)

}
\references{
Kim, Y. S. (2020) Portfolio Optimization on the Dispersion Risk and the Asymmetric Tail Risk
\url{https://arxiv.org/pdf/2007.13972.pdf}
}
