% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNTS.R
\name{VaRetnts}
\alias{VaRetnts}
\title{VaRetnts}
\usage{
VaRetnts(eps, ntsparam)
}
\arguments{
\item{eps}{the significant level for CVaR. Real value between 0 and 1.}

\item{ntsparam}{A vector of the NTS parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
A vector of the standard NTS parameters \eqn{(\alpha, \theta, \beta)}.}
}
\value{
Value at Return of the NTS distribution.
}
\description{
\code{VaRetnts} calculates Conditional Value at Risk (CVaR, or expected shortfall ES)
of the NTS market model with parameters \eqn{(\alpha, \theta, \beta, \gamma, \mu)}.
If only three parameters are given, it calculates CVaR
of the standard NTS distribution with parameter \eqn{(\alpha, \theta, \beta)}
}
\examples{
library("temStaR")
alpha <- 1.2
theta <- 1
beta <- -0.2
ntsparam <- c(alpha, theta, beta)
u <- c(0.01,0.05)
q <- VaRetnts(u, ntsparam)

alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
ntsparam <- c(alpha, theta, beta, gamma, mu)
u <- c(0.01,0.05)
q <- VaRetnts(u, ntsparam)


#Annual based parameters
alpha <- 1.2
theta <- 1
beta <- -0.2
gamma <- 0.3
mu <- 0.1
#scaling annual parameters to one day
dt <- 1/250 #one day
ntsparam <- c(alpha, theta, beta, gamma, mu, dt)
u <- c(0.01,0.05)
q <- VaRetnts(u, ntsparam)

}
\references{
Y. S. Kim, S. T. Rachev, M. L. Bianchi, and F. J. Fabozzi (2010), Computing VaR and AVaR in infinitely divisible distributions, Probability and Mathematical Statistics, 30 (2), 223-245.

S. T. Rachev, Y. S. Kim, M. L. Bianchi, and F. J. Fabozzi (2011), Financial Models with Levy Processes and Volatility Clustering, John Wiley & Sons
}
