% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMultiNTS.R
\name{qmarginalmnts}
\alias{qmarginalmnts}
\title{qmarginalmnts}
\usage{
qmarginalmnts(u, n, st)
}
\arguments{
\item{u}{vector of probabilities.}

\item{n}{the \eqn{n}-th element to be calculated.}

\item{st}{Structure of parameters for the n-dimensional NTS distribution.}
}
\description{
\code{qmarginalmnts} calculates
the quantile value of the \eqn{n}-th element
of the multivariate NTS distributed random variable.
}
